# packages/shared/spec/openapi.yaml
openapi: 3.0.3
info:
  title: artifacts_mmo server
  version: 0.1.0
components:
  schemas:
    Item:
      type: object
      required:
        - name
        - level
        - description
      properties:
        name:
          type: string
        level:
          type: number
        description:
          type: string
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: number
        detail:
          type: string
        upstreamCode:
          type: number
    HealthStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
paths:
  /items/:
    get:
      parameters:
        - schema:
            anyOf:
              - type: string
                enum:
                  - weaponcrafting
              - type: string
                enum:
                  - gearcrafting
              - type: string
                enum:
                  - jewelrycrafting
              - type: string
                enum:
                  - cooking
              - type: string
                enum:
                  - woodcutting
              - type: string
                enum:
                  - mining
              - type: string
                enum:
                  - alchemy
          in: query
          name: skill
          required: false
        - schema:
            type: number
          in: query
          name: level
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "500":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /health:
    get:
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /server-status:
    get:
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema: {}
        "500":
          description: Default Response
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /test:
    get:
      responses:
        "200":
          description: Default Response
  /fight-chicken:
    get:
      responses:
        "200":
          description: Default Response
  /error:
    get:
      responses:
        "200":
          description: Default Response