# packages/webapp/generated-src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type Item = {
    name: string;
    level: number;
    description: string;
};

export type Problem = {
    type?: string;
    title: string;
    status: number;
    detail?: string;
    upstreamCode?: number;
};

export type HealthStatus = {
    status: 'ok';
};

export type GetItemsData = {
    body?: never;
    path?: never;
    query?: {
        skill?: 'weaponcrafting' | 'gearcrafting' | 'jewelrycrafting' | 'cooking' | 'woodcutting' | 'mining' | 'alchemy';
        level?: number;
    };
    url: '/items/';
};

export type GetItemsErrors = {
    /**
     * Default Response
     */
    500: Problem;
};

export type GetItemsError = GetItemsErrors[keyof GetItemsErrors];

export type GetItemsResponses = {
    /**
     * Default Response
     */
    200: Array<Item>;
};

export type GetItemsResponse = GetItemsResponses[keyof GetItemsResponses];

export type GetHealthData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health';
};

export type GetHealthResponses = {
    /**
     * Default Response
     */
    200: HealthStatus;
};

export type GetHealthResponse = GetHealthResponses[keyof GetHealthResponses];

export type GetServerStatusData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/server-status';
};

export type GetServerStatusErrors = {
    /**
     * Default Response
     */
    500: Problem;
};

export type GetServerStatusError = GetServerStatusErrors[keyof GetServerStatusErrors];

export type GetServerStatusResponses = {
    /**
     * Default Response
     */
    200: unknown;
};

export type GetTestData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/test';
};

export type GetTestResponses = {
    /**
     * Default Response
     */
    200: unknown;
};

export type GetFightChickenData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/fight-chicken';
};

export type GetFightChickenResponses = {
    /**
     * Default Response
     */
    200: unknown;
};

export type GetErrorData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/error';
};

export type GetErrorResponses = {
    /**
     * Default Response
     */
    200: unknown;
};